backend/requirements.txt
fastapi
uvicorn
torch
torchvision
faiss-cpu
pillow
numpy
python-multipart